Extract jackson dependency from finance-workflows and move it to IRS Demo where it is actually needed
