+++ finance/workflows/build.gradle
@@ -31,9 +31,6 @@ dependencies {
     
     cordapp project(':finance:contracts')
 
-    // For JSON
-    compile "com.fasterxml.jackson.core:jackson-databind:${jackson_version}"
-    
     testCompile project(':test-utils')
     testCompile project(path: ':core', configuration: 'testArtifacts')
