+++ samples/irs-demo/cordapp/contracts-irs/build.gradle
@@ -9,6 +9,7 @@ dependencies {
     cordapp project(':finance:contracts')
 
     compile "com.fasterxml.jackson.core:jackson-annotations:${jackson_version}"
+    compile("com.fasterxml.jackson.module:jackson-module-kotlin:$jackson_version")
 
     testCompile project(':node-driver')
     testImplementation "org.junit.jupiter:junit-jupiter-api:${junit_jupiter_version}"
