+++ samples/irs-demo/cordapp/contracts-irs/src/main/kotlin/net/corda/irs/utilities/FinanceJSONSupport.kt
@@ -1,15 +1,13 @@
 @file:JvmName("FinanceJSONSupport")
 
-package net.corda.finance.workflows.plugin
+package net.corda.irs.utilities
 
-import TEST_CALENDAR_NAMES
 import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.core.JsonParseException
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.*
 import com.fasterxml.jackson.databind.deser.std.StringArrayDeserializer
 import com.fasterxml.jackson.databind.module.SimpleModule
-import loadTestCalendar
 import net.corda.finance.contracts.BusinessCalendar
 import net.corda.finance.contracts.Expression
 import java.time.LocalDate
