+++ samples/irs-demo/cordapp/contracts-irs/src/main/kotlin/net/corda/irs/utilities/FinanceWorkflowsUtils.kt
@@ -1,3 +1,5 @@
+package net.corda.irs.utilities
+
 import net.corda.core.flows.FlowException
 import net.corda.core.serialization.CordaSerializable
 import net.corda.finance.contracts.BusinessCalendar
@@ -5,7 +7,7 @@ import net.corda.finance.contracts.BusinessCalendar
 val TEST_CALENDAR_NAMES = listOf("London", "NewYork")
 
 fun loadTestCalendar(name: String): BusinessCalendar {
-    val stream = UnknownCalendar::class.java.getResourceAsStream("net/corda/finance/workflows/utils/${name}HolidayCalendar.txt") ?: throw UnknownCalendar(name)
+    val stream = UnknownCalendar::class.java.getResourceAsStream("/net/corda/finance/workflows/utils/${name}HolidayCalendar.txt") ?: throw UnknownCalendar(name)
     return stream.use {
         BusinessCalendar(stream.reader().readText().split(",").map { BusinessCalendar.parseDateFromString(it) }.toSortedSet())
     }
