+++ samples/irs-demo/cordapp/contracts-irs/src/main/resources/net/corda/finance/workflows/utils/NewYorkHolidayCalendar.txt
